+++ core/src/db/DB.h
@@ -80,7 +80,7 @@ class DB {
     DropPartitionByTag(const std::string& table_id, const std::string& partition_tag) = 0;
 
     virtual Status
-    ShowPartitions(const std::string& table_id, std::vector<meta::TableSchema>& partiton_schema_array) = 0;
+    ShowPartitions(const std::string& table_id, std::vector<meta::TableSchema>& partition_schema_array) = 0;
 
     virtual Status
     InsertVectors(const std::string& table_id, const std::string& partition_tag, uint64_t n, const float* vectors,